Changed implementation to use integers instead of strings
The internal implementation now uses Java 32 bit integers to represent
the 16 bit data instead of strings, which is much more wasteful. While
it would be nicer to use shorts, the Java short data type is only
signed, which would require more casting to work in processing code.

Externally, string input/output is still supported, but not recommend
for processing use.